Improve KilimContinuationRingBenchmark and fix RingBenchmarkTestUtil#verifyResult().
verifyResult() change is backwards incompatible. Now every test is
failing except KilimContinuationRingBenchmarkTest. But this is for good.
Previously we were performing an extra WORKER_COUNT amount of message
passings. This is due to the legacy "start-run-stop" behaviour employed
in the benchmarks, whereas I want to switch to
"start-run-run-run-...-run-stop" scheme.

By the way, SLF4J and Logback dependencies are dropped as well.